Ensure window.BeyondWords is available when running ./bin/server
When ./bin/server script is running, it uses a different module loading
approach than when a dist/umd.js script tag is added to the page. A
consequence of this is that exports from src/index.ts are not available
as globals on the window object. Therefore, make this behaviour
consistent by explicitly setting window.BeyondWords in that context.