Return a svelte component from `new BeyondWords.Player({ … })`
This makes it so the top-level instance returned to the SDK user is a
svelte component so that they can make use of its reactivity, e.g.

```
player.userInterface = false;
```

Otherwise, we’d have to provide an explicit method for each operation we
want to support which would most likely add a lot of duplication.