Refactor to move the time component out of the ProgressBar
In the ‘icon’ style player, the ‘Listen to this article’ text remains
while the podcast is playing. Therefore, I think it makes sense to try
and reuse the existing time component that was provided to the
ProgressBar slot, rather than duplicate it in the UserInterface
component. To reuse it, we need it to be a top-level component of the
UserInterface rather than nested inside ProgressBar so that means having
to re-do quite a few styles.