Choose which player to show as the widget at the bottom of the page
For a player to be eligible, the user must have scrolled past the player
such that it is not visible on screen. It then chooses based on which
players are playing/paused and if none are, it chooses the lowest player
on the page - the one that was most recently scrolled past.