Add the PodcastTitle to the standard player above 700px width
Make its width match the design for the 1200px width example. If the
title doesn't fit, it shows elipses on the end. If the full title is
shown, it does not grow any more and the progress bar grows instead.

The value 0.52 is specifically to match the designs but is TBC.