Use css visibility to control whether PlayerTitle shows
I want to expand/collapse the small player widget with an animation but
this requires that the width of the expanded/collapsed elements are
specified in css for the transition animation to work. I plan to measure
the width of these elements with JavaScript onMount and then set the css
rule. I therefore want to make sure that the width is always set to a
positive value onMount which wouldn’t have been the case if an advert is
playing because playerTitle was set to “”. Therefore, use visibility to
hide the element so that the width is always calculated by the browser.