Style the high-level DOM structure for the video player
Add a video element to the .hoverable div and set it to 100% of its size.
The hoverable div uses padding-bottom to enforce a 16:9 aspect ratio.
The main div then ignores this padding so that it is overlaid on top of
the video. We use a grid layout to position .controls / .end elements.