Make calculating the width of the PlayerTitle more robust
The non-widget small player was being truncated to ‘Listen to this artic’
because the width calculation seems to change very soon after the
component is added to the DOM - possibly caused by font loading.

To make this more robust, add a delay of 10ms and then repeat the
calculation a few times after the component has loaded in case the fonts
take a while to load. We use scrollWidth because this is not dependent
on the current width of the component.