Make the playwright tests consistent wherever they are run
The playwright tests take screenshots of real browsers which means there
might be subtle differences depending on which platform the tests are
run on. For example, font rendering might differ slightly in Chrome on
Windows, macOS and Ubuntu.

We need the screenshots to be the same every time because otherwise the
visual regression tests will fail and give a false negative.

In GitHub actions, we can run the entire build inside the playwright
Docker image so that we don’t need to run `npm install` twice.

We also no longer need to install the playwright dependencies in the
./bin/setup script because, by default, the playwright tests run against
the Docker image. For convenience the ./bin/test_features script takes
a flag that allows the tests to run natively on the local machine, e.g.

```
./bin/test_features --native --headed
```