Disable screenshot tests for Edge and update to iPhone 14 / Pixel 7
Microsoft Edge isn’t currently supported by playwright in linux so we
are no longer able to test against this browser now that tests are run
in the Docker container.

The iPhone 12 device was sometimes hanging for an unknown reasons so
upgrade to the latest version which does not seem to have this problem.
Upgrade to the latest Pixel version for consistency.