Isolate node_modules/ used by the playwright Docker container
Previously, when ./bin/test_features ran it was doing a clean-install of
node_modules within the Docker container. The Docker container uses
ubuntu and the packages are not compatible with both macOS and linux.
This meant that run ./bin/test_units afterwards would fail and a clean
install would be required on the host machine.

Instead, mount node_modules/ within Docker to the node_modules/.test
directory so that it does not interfere.