Override the active style to ‘video’ when maximized
If the interfaceStyle is set to something other than ‘video’ and the
widgetStyle is set to ‘video’, then clicking maximize in the widget will
always show the non-widget interface because making it fullscreen
scrolls that interface into view. Therefore, we need to temporarily
override the interfaceStyle to ‘video’ so that the video actually shows
in fullscreen, otherwise we’d just show the audio player.