Fix the screenshot tests
Previously we were setting the video poster attribute which was always
showing in the screenshot tests because we don’t set any media. Also,
the poster attribute seems a bit unreliable in desktop/mobile Safari
which sometimes was not clearing the poster image after the attribute
was removed, causing some tests to intermittently fail.

Instead, use the .video-placeholder to set the poster image in
UserInterface. The logic seems like it belongs there anyway.