Ensure the HLS stream is attached with the same playback state/rate
playbackRate definitely isn’t being applied after re-attaching a HLS
source so we need to set it manually. The playback state is being
applied but I think this only works because of the execution order of
the reactive expressions in MediaElement, so I’d rather confirm that it
is set correctly here, too.