Don’t re-assign media if contentIndex changes while an ad is playing
Previously, the value of currentAdvert.media was re-assigned to media
whenever contentIndex changed, even when activeAdvert was true. This was
causing Svelte to trigger reactive changes to media even though it
hadn’t change. This had the side effect that media would be reloaded
when changing playlist items when an ad is playing - even though the
media hadn’t actually changed (it was still the ad).