Keep track of playedAdvertMedia when choosing an advert
We don’t want to play the same advert media more than once, so keep a
global set of played advert media and skip adverts that have already
been played. I think it’s reasonable to look for .some() because all the
media should be the same - just in different formats.