Don’t play more than one advert in a row
After a pre-roll advert ends, the time is set back to zero for the
content. The timeupdate then calls #chooseAndSetAdvert again and because
the time is zero, the content is again eligible for a pre-roll advert.
This causes all of the adverts to be played in sequence one after
another rather than interspersed between content.

Therefore, set a ‘skipNextAdvert’ flag in this situation so that
#chooseAndSetAdvert knows that this situation has arisen. It then clears
the flag so that future adverts can play.

I think this will also mean that if a post-roll ad plays and the content
finishes, if the user clicks on play again then will not receive an
advert because the flag is set. I think this is fine because these ads
are still technically being played in sequence - just not immediately
after one another.