Fix showing the wrong duration for content if an advert is about to play
If an advert is about to play but currentTime is 0 and the player is
paused, we show the contentItem (not the advert) in the user interface
so as not to discourage users from clicking on the player. However, the
duration is shown when paused so we need to make sure we get the
duration from the contentItem, rather than relying on the duration of
audio loaded into the media element.