Fix a bug with scrubbing and playback restarting
If you set currentTime equal to duration while paused, the video element
restarts playback from currentTime=0 for some reason. I’m not sure if
this is a bug in Svelte or something odd about the HTML video
specification for what should happen in that scenario.

We can avoid this by ensuring currentTime is always less than duration.