Defer setting advertIndex until the player is playing
If a playlist is paused and then the user clicks on the next track, we
don’t want to set the advert until the player is playing.

Previously, I worked around this in UserInterface by hiding the fact
that an advert was set, but that approach also caused other
inconsistencies, such as not being able to scrub the progress bar
because an advert technically set, even though it wasn’t shown.