Remove the TODO about ‘contentComplete’
This signals to Google that we can play post-roll ads, but we control
this process ourselves so don’t need to send this signal. I think this
means we’ll need to make sure that all programmatic ads are set up with
a “pre-roll” (or “mid-roll” ?) placement to be served to the player.