Make it so you can click on a playing video to open advert links
To make this work, we need to disable pointer events on the user
interface so that the clicks go through to the Google iframe. For
consistency, we also add a full-size anchor to the MediaElement for
custom adverts so that it behaves in the same way.

A consequence of disabling pointer events is that nothing responds to
mouse clicks in the user interface, which isn’t what we we want.
Therefore, we re-enable pointer events on the .controls and .end (logo)
elements so that they continue to work.

It also means that the Hoverable component no longer receives mouse
events which means the background overlay doesn’t show, but this is
probably what we want for adverts, anyway.