Hide the AdvertButton when there’s no href
The Google IMA SDK doesn’t provide a supported method for retrieving
the clickThroughUrl. It is technically possible [0] but using this
approach is unsupported and I can see that it is not firing the tracking
events which are presumably used to determine how successful an ad
campaign is. Therefore, I think it’s better to hide the AdvertButton and
prefer users to click on the playing video to visit the advert link,
rather than trying to provide one of our own for VAST adverts.

[0] https://stackoverflow.com/questions/34782581/how-to-trigger-html5-ima-sdk-click-event-manually