Remove a TODO about loading the ima3.js script sooner
I think it’s fine for now to load the script when the first VAST advert
is due to play. The script is ~120KB and might already be cached in the
user’s browser. There’s already some expectation that there will be a
delay given they are trying to play an audio file that would likely need
to be fetched.

An alternative approach would be load the ima3.js script after the
player loads if it detects there are VAST adverts and there is content
with adsEnabled, but this would mean the script loads on lots of pages
where they user never clicks play which seems wasteful.