Reduce the set of valid analytics event types
These are the only event types that are actually used by the backend:

- load: when the player is first loaded on the page
- play: when playback starts within the player
- play_progress: when a 10% time increment is reached
- ad_link_click: when the user clicks on the advert

It also uses ‘ad_logo_click’ but for the same as ‘ad_logo_click’ so we
can simplify it and just emit one. It’s not clear what this event type
would map to in the new player, anyway.

The backend uses a uniqBy(listen_session_id, …) to avoid counting
duplicates of the ‘play’ event.