Fix two ‘PlaybackPlaying’ events being emitted
Previously, we’d set playbackState = “playing” and then wait for a
<video onplay=“…”> event. We’d then try to play the deferred advert if
there is one. This would cause playback to temporarily begin and then
switch to the advert once the second Svelte update occurs.

A consequence of this is that two PlaybackPlaying events were emitted
and you’d sometimes hear a brief moment of audio from the content before
the advert begins which is wrong.