Add a GoogleAnalytics component that loads the external library
We don’t need to load the library again if gtag is already defined. It
doesn’t matter if it was loaded with a different Measurement ID because
we will use ‘send_to: …’ in the events to send it to our tag only.

We don’t call `gtag('config', analyticsTag);` because we don’t want to
send other page view events, etc. We only want to send custom events to
this measurement ID.