Capture errors into Sentry if they came from the player
The player will be embedded on publishers’ websites so we don’t want to
capture every error because it would include errors that are nothing to
do with the player. Therefore, check if the error originated from the
compiled script (that will also contain initializeSentry).

On development, the script will still be called initializeSentry so we
can use this to determine the environment. We don’t currently have any
notion of a ‘staging’ environment for the player.