Add a ./bin/upload_sourcemap script to upload to Sentry
This un-minifies trace information for errors reported to Sentry which
is useful for debugging. The script does not run ./bin/build which must
be run to produce the sourcemap file. The script will not upload the
sourcemap if a release already exists with that version. I intend to use
the script as part of a ./bin/publish script which will ensure the
version is incremented.