Tweak the controller/player load order
Previously, the player was initialised and then the controller prop was
set on it. This meant that any reactive properties that used the
controller were being triggered twice - once for the initial load with
an undefined value and then again when controller was set.

Instead, set the controller once in the player initialiser but use async
to wait for the player to be set before processing any events. This is
neater and means that onStatusChange isn’t registering multiple times.