Fix the TimeIndicator briefly fading out after an advert
During an advert, the TimeIndicator is always visible whereas the
controls are hidden if the user is not interacting with the player.

When the advert ends, the TimeIndicator switches back to its normal
position next to the play button but it has a ‘transition: opacity 0.25s’
rule that means we briefly see it in its new position then immediately
fade out since the advert is no longer player.

To fix this glitch, we can remove the opacity transition rule briefly
when content begins playing if it is not an advert.