Add a chooseMediaSession helper
The Media Session API can be used to allow the player to be controlled
on the iPhone lock screen and in other places, depending on the device.

There is one global navigator.mediaSession for the page so we need to
be careful not to override the publisher’s media session if they are
setting this themselves and we need to choose which player should be
controlled by the media session.

The logic for this is similar to choosing a widget since we want to
choose zero or one of our players to show in the media session, so I’ve
tried to write the code in the same manner for consistency.