Add support for the ‘sizes: “any”’ MediaSession artwork
The W3C specification states that you can set sizes to ‘any’ to use the
provided image as the artwork. This doesn’t seem to work on iOS but it
might in devices that follow the spec more closely. Add this artwork
last so that it does not break on iOS.