Add more MediaSession handler logic…
Choose whether to register the ‘seek’ handlers or the ‘track’ handlers
based on the skip button style. The device will check which handlers are
registered and then decide which buttons to show.

If an advert is playing, don’t register handlers so that the user can’t
skip forwards or backwards using the button. We already ignore these
events in RootController while an advert is playing but this approach
will also disable the device buttons.

Emit the same events but change initiatedBy so we can distinguish these
events.