Handle ‘seekto’ events in the Media Session
These events are emitted once when the user releases their finger from
the progress bar on the iOS lock screen. We don’t need to handle the
started/finished scrubbing logic because it’s done for us by the device.

While an advert is playing, scrubbing is prevented. This alongside the
disabling of the skip buttons means the user must listen to the advert
which is consistent with the behaviour of the player on the webpage.