Make the style rules extremely defensive against accidental overrides
Prefix all styles within the player with “.bwp”.repeat(12). That means
a style rule with a specificity of at least 15 would be required to
override the player styles, e.g.

.bwp.bwp.bwp.bwp.bwp.bwp.bwp.bwp.bwp.bwp.bwp.bwp.bwp.bwp.bwp .main {
  border-radius: 0;
}

It can still be done by someone who really wants to but this should
ensure that no accidental overrides occur.