Return the postcss sourcemap from the vite plugin
There was a build warning about the plugin not generating a sourcemap.
We could set it to null to ignore the sourcemap but postcss seems to
provide one - although I’m not sure if it’s actually used for CSS
output. The resultant CSS is then fed into the JavaScript but I don’t
think the sourcemap will have any bearing on that.