Make segmentsPlayback work with adverts
Set prevContent and prevTime if a segment is clicked on while an
advert is playing. This shouldn’t clamp to the duration because that is
the duration of the advert and not the content.

If the track is changed, the segment resets to -1 and prevTime is also
set to 0 so that it starts from the beginning. We didn’t need to do this
before because you couldn’t start a different track mid-way through
whereas now you can by clicking on the segment.