Simplify how segmentIndex is set
Previously, the Player component also set the segmentIndex in response
to currentTime changes but this made it confusing since the controller
also set this property. The reason it was done this way is because
currentTime updates for more frequently than the CurrentTimeUpdated
event because Svelte uses requestAnimationFrame to make it smoother,
which makes for a more accurate change of highlighting.

This new approach makes segmentIndex entirely owned by the controller,
but emits new CurrentSegmentUpdated events that use the more precise
currentTime property so that the highlighting is more accurate.