Don’t highlight the first segment while the advert plays
If the user clicks on the play button (not the segment), the first
segment was highlighted because the currentTime is set to 0 before
the activeAdvert state is set. It then remains highlight for the
duration of the advert. I think this is a bit weird so only show
highlights during adverts if the user explicitly clicked on a segment.