Make all inline CSS in the player important
There really isn’t an easy way to do this. I tried preprocessing the
source code of Svelte components to modify the style=“…” attributes but
this turned out to be complicated because style can be set to arbitrary
expressions.

This approach is coupled to Svelte’s function names but seems to work,
although it doesn’t apply !important to other styles that Svelte adds
itself, e.g. when using animations.