Fix the MediaElement’s fly-widget animation
Browsers do not support setting !important on @keyframe rules so the
!important in the non-keyframe rules always take precedence, regardless
of specificity. This broke the animation.

To fix it, this introduced a new .animating class that is temporarily
added for 0.5 seconds when showBehindWidget becomes true. When this
class is set, the ‘all: initial !important’ rule is disabled and the
rules overridden by the animation are disabled by using the exclusion
comment /* ~!important */.

The downside of this approach is that during that 0.5 seconds, the
publisher could override the bottom/opacity styles if they set
!important on their CSS rules - but this seems very unlikely.

I spent a very long time thinking about this and I think this is the
only way to get this animation to work if we want to add !important to
every style rule to be extra defensive against publisher overrides.