Document that !important is needed to override styles
You can start to override some of the styles at precedence 12 but need
to go to precedence 15 override all of them, e.g. border-radius on .main

Now that we have !important on all our styles, I think it’s safe to go
back down to 15 instead of requiring 18 classes to override styles.