Use a better strategy when highlighting segments
Previously, the highlightColor would be applied from the first player
that highlighted a segment and this would remain until the reference
count reached zero. If a player set a highlightColor and then ownership
changed to a different player, the wrong highlightColor would continue
to show.

More generally, I need to use an ownership pattern like this for the
segment widget which should belong to one player at a time, but could
revert to other players when the segment of the player changes.