Add a SegmentUIContainers class
Use the same OwnershipMediator to ensure that each segment element is
owned by at most one player instance. This is a first attempt that does
not yet add/remove container elements adjacent to the markers. It just
returns the markers themselves. These elements are not yet used to mount
the component tree and are ignored in the Svelte #each loop which means
the element just appears directly below the existing user interface.