Extract the ‘clickable’ responsibility from SegmentHighlights
Previously, this class was responsible for two things:

1) Adding/removing mark elements to highlight text within segments
2) Adding/removing the .beyondwords-clickable class to the segment

The class added in 2) makes it so the ‘cursor: pointer’ shows over the
segment to indicate it is clickable. As stated in an earlier commit,
it’s possible to highlight segments without them being clickable and
vice versa, so we should probably separate these responsibilities.

Further, the player that owns whether a segment is clickable might not
be the same as the one that is highlighting it. For example, if a
segment is clickable in one player but not another, then when the
segment is highlighted by the other player, it would have become not
clickable (no cursor: pointer) even though it can be clicked on to begin
playback in the first player.