Remove the segment widgets if an advert plays when no segment was pressed
We already make it so we don’t highlight the first segment when the
advert is playing if no segment has been clicked. We should do a similar
things for the segment widgets and hide them if the user clicks play on
the inline-player without explicitly clicking on a segment (or its widget).