Fix the segment widget being removed when it shouldn’t
Previously, we were clearing lastHovered so that currentSegment can take
over from it in the segment widget but we shouldn’t do that if the
current segment is not allowed to show in the widget because of its
section. This fixes the segment widget hiding after the user hovers over
a segment and then the current segment changes.