Emit a VisibilityChanged event when showUserInterface changes
If showUserInterface is toggled the player’s visibility might change.
Previously, we only detected cases when the player was scrolled into or
out of view but we should probably also choose a new widget if you
disable the user-interface, for example, since this is also one of the
criteria we check in chooseWidget.