Re-work how the segment widget is mounted
I want to use the same mechanism for mounting the segment widget in an
arbitrary location on the page for the bottom-of-page widget via an
optional player setting (widgetTarget). This pre-work makes the
code a bit more uniform and re-uses style rules so that the widgetTarget
setting can be added.