Add a widgetTarget setting to allow the widget to be mounted anywhere
In the BeyondWords dashboard, we want to mount the widget inline in the
page, rather than have it show in the bottom corner of the screen. To
support this feature, introduce a widgetTarget setting that uses the
same mechanism to externally mount elements outside of the player’s
component tree.

To keep things simpler, if a widgetTarget is set then the widget will
always show and the player will be ignored when choosing a widget to
show at the bottom of the screen. It is up to the integrator if they
want to show and hide the widget.

Additionally, the current behaviour of hiding the widget when
showUserInterface is disabled does not affect the widget if a
widgetTarget is set. Again, the integrator can remove the widgetTarget
(reverting it to show at the bottom of screen) and then set
showUserInterface to false to hide it, if they need to.