Fix VAST advert playback becoming stuck on iOS
For some reason, on iOS, playing a VAST advert while setting the
currentTime of the video tag causes the advert to become stuck. I think
this is because of a difference in how the video tag is used to play the
media on iOS, i.e. by setting the video.src attribute which doesn’t seem
to occur in other browsers when using the web inspector.

You probably shouldn’t be able to adjust the currentTime of the player
through the SDK when adverts are playing anyway since this is not
allowed in the user interface, e.g. the progress bar is disabled.