Simplify the prefixCssSelectors logic
Previously, we needed to set lower specificity on the ‘all: initial’
reset rule because the :not selector increases its specificity for each
additional term. We can avoid this by using the :where CSS pseudo class
which has a specificity of 0.

This means that the plugins in vite.config.ts can be removed and the
player continues to render correctly. Previously, the three-levels of
specificity in the :not clause meant that you could not remove these
lines without breaking the layout, which is pretty confusing.