Make the ProgressBar extend to the end of the UI when a VAST advert is playing
We don’t show the AdvertButton when a VAST advert is playing because the
user can click on the video background.

Audio-only VAST adverts also don’t seem to provide a URL that can be
clicked on. I tried setting clickTrackingElement but that had no effect:
https://developers.google.com/interactive-media-ads/docs/sdks/html5/client-side/reference/js/google.ima.AdDisplayContainer

Presumably, audio-only adverts are intended to be used when no interface
is available, e.g. when listening to your phone through headphones.