Fix the widgetTarget not being respected
The ExternalWidget component moves its slot into the widgetTarget but
there was a bug where the widgetTarget would be ignored if Svelte
re-mounted the content of within the slot, e.g. by setting widgetStyle
to “none” and then to “standard” again.

Instead, we should not move the slot directly but move the outer div so
that the content that has been moved is never re-mounted.

We should also set ‘position: relative’ on the div so that absolutely
positioned elements, e.g. TimeIndicator are relative to the container
and not the page.