Fix the screen player hiding during an advert when AdBlock is enabled
When using the AdBlock extension, if an advert is played on the mobile
sized screen player, the entire player disappears for the duration of
the advert. I think it is detecting elements in the DOM that have a
class of .advert and hiding them, but this completely breaks the
experience of the player and the advert plays anyway.

Therefore, rename the advert CSS classes throughout the app so that
AdBlock does not interfere with element styles.