Drop support for preventing some attributes from being overridden
This was a nice idea but didn’t really materialise. We could pass flags
through from the backend to determine which fields can be overridden but
1) this adds complexity and 2) there’s nothing to stop integrators
overriding these fields using the player SDK (after initialization).

In the short term, we need to set ‘logoIconEnabled: true’ on the
standalone /play and /playlist pages in the backend so it’s probably
best to remove this logic for now.