Start working on the controller logic for intros/outros
Model the logic after adverts which are pretty similar in that they play
before/after content and are set when the identifiers change and
in-between each track.

I think the behaviour should be that the intro plays after the advert
and before the content, even if it’s not the first content item, e.g. if
the user clicked on the second item in a playlist.

While the intro/outro is playing, the user should be able to change the
playlist track and the segment without affecting playback (like adverts).