Skip the intro if the user is not at currentTime=0
If the user has clicked on a segment or scrubbed the large player’s
progress bar forwards before clicking play, we should skip the intro.
We need to perform the check of whether a pre-roll advert should play
again in this case because normally it would have been set as a deferred
advert, but the intro was set as deferred instead.