Fix content starting from the beginning in the large player
If you scrub the progress bar while stopped and then press play, it was
starting playback from the beginning of the content, rather than at the
scrubbed position. This was because we were temporarily setting the
introsOutrosIndex on the player which was changing the media which
results in events being emitted that reset the time to 0.

Rather than set introsOutrosIndex then check if we should skip the
intro, perform this check first to avoid temporarily setting the
incorrect index on the player.