Fix the currentSegment not being highlighted while the advert plays
If the user scrubs or presses on the progress bar in the large player
while the content is stopped, the currentSegment is highlighted based
on the currentTime of the player. However, if an advert then plays, we
only continue to highlight the currentSegment if segmentPlayed has been
set. We probably don’t want to do this if they scrub all the way back to
time=0 because then the intro plays first.