Show the real playback time in the segment widget during intro/outro
If the intro or outro is playing, we should show the real time the
segment widget and not the startTime for that segment. If we did try to
show the startTime, the progress circle would not be correct because the
duration is currently set to that of the intro/outro and not the content.