Fix transitions not undoing in the built umd.js file
settableProps was returning an empty object every time because Svelte
does not expose the settable properties of a component int he production
build. In the short term, I’ve fixed this by using a different way to
fetch the keys: Object.keys(object.$$props) but this probably belongs on
the top-level player component in index.ts.