Don’t test .animating elements for accessibility issues
When ./bin/test_features is running in --headed mode, the widget
animation plays where it fades in from the bottom of the screen. If the
accessibility check runs mid-way through this animation it might fail
the standard because of low contrast.

In this case, ignore the .animating element. Note that this shouldn’t
affect the headless tests that run on CI because the CSS animations
don’t play in a headless context.