Fix the video advert UI becoming stuck when clicking external links
If an external link is clicked when an advert is playing in the video
playerStyle, the element that was clicked remains focussed when the user
returns to the tab which causes the :focus-visible styles to remain in
effect which means that video overlay is shown when the user isn’t
hovering over it which is incorrect.

To fix this, blur all target=“_blank” links on mouse.