Use well established browser-supported features to set the initial time
Previously, we were trying to set the initial time of the player based
on events but that was unreliably because browsers emitted events in
subtly different orders.

Instead, use a media fragments[1] to set the initial time for sources
that can play natively and set the startPosition prop for sources that
play through the HLS library.