Ensure the UI is correct when calling play/pause on the underlying video tag
Annoyingly, in some cases, play() and pause() are called on the
underlying <video> HTML element rather than going through the public
SDK provided by the player. For example, on Android, when receiving a
phone call, the pause() method is called on all HTML media elements.

The player wasn’t really built with this use case in mind since it is
supposed to abstract over the <video> so its playbackState does not
automatically update when these events happen. Furthermore, it’s quite
difficult to isolate these events from conventional events emitted by
the player, e.g. PlaybackPaused is emitted just before PlaybackEnded
when a track ends.

I think these conditionals do the trick and ensure that the UI is kept
up to date when calling play/pause on the underlying <video> element.