Dynamically import the HLS library to reduce the player package size
The HLS library isn’t needed in Safari which supports native HLS
playback and it won’t be needed when playing other formats, e.g. mp3.

Eventually, we might want to loadMedia once the user clicks play in
which case we can defer this network request until then.

In the previous approach, the HLS library was bundled into the player
umd.js file so none of the above was an option.