Fix error stacktraces not pointing to the relevant line
One of the plugins makes all inline css important and it does so by
modifying the source code emitted by Svelte during build time. We need
to ensure we generate a source map for this phase otherwise any errors
logged to the JavaScript console will point to the wrong location.