Fix the player pausing when touching on the progress bar
When you touch on the progress bar, the browser sends a touchstart and
touchend event very close together. This creates a race condition where
the PlaybackPaused event is emitted after we have already transitioned
the player back to its preScrubState (playing) which was causing it to
become paused again. We can fix this by introducing a short delay before
clearing preScrubState which shouldn’t cause any problems since this
state is overridden when we start scrubbing again, anyway.