Fix a CORS issue when importing in umd.js
The import statements in umd.js look like this:
import(“./././style.js”)

But this fails to load with an error:
> The base URL is about:blank because import() is called from a CORS-cross-origin script.

Unfortunately, we can’t use relative paths in the umd.js file.
Previously it worked because localhost is whitelisted so the issue
wasn’t found until released to NPM.

Therefore, introduce some utility methods on BeyondWords.Player to get
the styleSrc and hlsSrc URLs (via stacktraceparser) and replace the
argument to import with the expanded URLs.

Alternatively, we could fix this by setting <script type=“module”> but
I’d rather avoid that since some publishers are already using the script
without it and I’d rather not introduce additional constraints.