Wait for MediaLoaded before sending the ‘load’ analytics event
The MediaLoaded event was recently added and I think it would be better
to use this rather than infer that the player has loaded if the duration
has been set. This also ensures the media_id and media_format fields are
available when the load event is sent, whereas previously they were sent
as null.