Fix an issue with the umd.js and RequireJS
When the umd.js script is loaded with an async tag and RequireJS is
already loaded on the page, it results in an error. This commit adds a
Vite plugin based on this one:

https://github.com/pakholeung37/vite-plugin-treat-umd-as-commonjs/blob/master/src/index.ts#L19

The NPM package above hasn’t been updated in a while and its
dependencies aren’t compatible with Svelte 4 and the plugin did not try
to keep the sourcemap intact. Therefore, recreate the plugin ourselves.