Fix some issues with dynamic dependencies and caching
If the script source is set to:
https://proxy.beyondwords.io/npm/@beyondwords/player@latest/dist/umd.js

Then it will try to load:
https://proxy.beyondwords.io/npm/@beyondwords/player@latest/dist/style.js

This is working correctly but the files might be cached in which case
the styles loaded by the browser might be for a previous player version.
When this happens, the randomly generated Svelte classes aren’t targeted
by the style rules which causes the player to appear in a broken state.

To fix detect when the URL contains player@latest and replace it with a
URL that matches the current version of the player, e.g.
https://proxy.beyondwords.io/npm/@beyondwords/player@0.0.0-beta47/dist/style.js

Also, add a warning if we are unable to detect the current URL. I
haven’t seen this happen yet.