Fix sometimes sending multiple MediaLoaded events
Now that the Hls library is fetched dynamically, there is a brief period
where no Hls support is available while the video sources are set:

<source src=“foo.m3u8”>
<source src=“foo.mp3”>

This causes the browser to load the first supported source, which is the
mp3, then, when the loadHlsIfNeeded promise resolves, it will load the
m3u8. This can cause two MediaLoaded events to trigger.

To fix this, send a “pending” signal back from loadMedia if we intend to
load Hls but the promise has not resolved yet. Then, avoid adding any
sources to the <video> tag if this signal is detected.