Add new video*Color properties
I spent a long time designing a nice solution that used a ColorThemes
object, only to realise it wouldn’t work because we need to show two
different themes at the same time for the inline-player and the widget.

This was my branch:
https://github.com/beyondwords-io/player/compare/main...color-themes

Instead, the simplest approach is probably to just add more top-level
fields to store the video color theme.