Fix Android Chrome not setting the initial time for HLS sources
Android Chrome reports that it has native support for HLS but it doesn’t
seem to support setting the initial time via fragment, e.g. #t=15

When I try to use the HLS library with Android Chrome, I get a manifest
error with no further explanation, so setting the initial time via the
HLS library also won’t work.

Therefore, if we detect that the browser is Android Chrome and we are
trying to set the initial time of the player, we should demote HLS
sources to the bottom of the list so that we prefer other sources that
do support time fragments, e.g. mp3 and mp4.

I don’t think we should completely remove HLS sources from the list
because they are playable, it would just be annoying not to set the
desired initial time of the player.