Fix the video source not loading in Safari
The problem was that calls to video.load() or video.pause() resulted in
an AbortError to be thrown during the implicit loading of the video
source. The browser implementation must have then decided to blacklist
the source and continue onto the next available playable source which
was an audio source.

I opened (and closed) an issue about this here which has a simpler repro:
https://github.com/sveltejs/svelte/issues/9052